rename : 컴포넌트 이름 대문자로 변경경
