+++ src/components/RegisterForm/getInfo.js
