+++ src/components/RegisterForm/walkTimeMinutes.js
