+++ src/components/common/InputField/index.js
@@ -1,7 +1,7 @@
 import createElement from "../../../utils/createElement/createElement";
-import Input from "../input";
-import Select from "../select";
-import TextArea from "../textArea";
+import Input from "../Input";
+import Select from "../Select";
+import TextArea from "../TextArea";
 import createFormSelectorByType from "./createFormSelectorByType";
 import HelpText from "./HelpText";
 import Label from "./Label";
