+++ src/components/common/Modal/index.js
@@ -1,5 +1,5 @@
-import BackDrop from "./backdrop";
-import ModalContent from "./modalContent";
+import BackDrop from "./Backdrop";
+import ModalContent from "./ModalContent";
 
 const Modal = (handleCloseModal, ...content) => {
   const backDrop = BackDrop(handleCloseModal);
