+++ src/components/common/Modal/registerModalClose.js
