+++ src/ui/renderRestaurant.js
@@ -1,4 +1,4 @@
-import RestaurantCard from "../components/restaurantCard";
+import RestaurantCard from "../components/RestaurantCard";
 import { $ } from "../utils/dom";
 
 const renderRestaurants = (...restaurantList) => {
